Add python script as entrypoint to Dockerfile
By adding the Python script to the entry point, running the container
will now execute the python script.

I also added `sys.argv` to the Python script to accept command line
arguments that can be used to parameterize the pipeline.

There are notes in the readme describing the docker commands